test: test for health endpoint.
